+++ client/src/pages/welcome/WelcomePage.tsx
@@ -1,6 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unused-expressions */
 import React from 'reactn'
 
+import welcomeDoctorImage from 'src/images/welcome-doctor.svg'
+
 import {
   Page,
   Block,
@@ -22,7 +24,9 @@ import { ReactNComponent } from 'reactn/build/components'
 import { NoCurrentUserError } from 'src/errors'
 
 import { toggleLocale, signOut } from 'src/initializers/providers'
-import { plural, Trans } from '@lingui/macro'
+import {
+  plural, Trans,
+} from '@lingui/macro'
 
 interface State {
   termsOpened: boolean
@@ -54,11 +58,8 @@ export default class WelcomePage extends ReactNComponent<any, State> {
     if (fillForSelf && fillForChildren) {
       return (
         <Trans id="WelcomePage.fill_for_self_and_children">
-          Every day you'll need to fill out symptom surveys for
-          you and your
-          {' '}
-          {this.global.i18n._(plural(user.children.length, { one: 'child', other: 'children' }))}
-          .
+          Every day you'll need to fill out symptom surveys for you and your{' '}
+          {plural(user.children.length, { one: 'child', other: 'children' })}.
         </Trans>
       )
     } if (fillForSelf) {
@@ -70,17 +71,16 @@ export default class WelcomePage extends ReactNComponent<any, State> {
     } if (fillForChildren) {
       return (
         <Trans id="WelcomePage.fill_children">
-          Every day you'll need to fill out symptom surveys for your
-          {this.global.i18n._(plural(user.children.length, { one: 'child', other: 'children' }))}
-          .
+          Every day you'll need to fill out symptom surveys for your{' '}
+          {plural(user.children.length, { one: 'child', other: 'children' })}.
         </Trans>
       )
     }
-    // TODO: What do we do in the case that the user is not associated with anything over the long run?
+
     return (
       <Trans id="WelcomePage.fill_for_no_one_error">
         It looks like your account has not been set up properly.
-        Please contact greenlight at help@greenlightready.com.
+        Please contact greenlight at <Link href="mailto:help@greenlightready.com">help@greenlightready.com</Link>.
       </Trans>
     )
   }
@@ -92,16 +92,13 @@ export default class WelcomePage extends ReactNComponent<any, State> {
       <Page>
         <Block>
           <h1>
-            {esExclaim()}
-            {greeting()}
-            ! &nbsp;&nbsp;&nbsp;&nbsp;
+            {esExclaim()}{greeting()}! &nbsp;&nbsp;&nbsp;&nbsp;
             <Link style={{ fontSize: '12px' }} onClick={() => toggleLocale()}>
               <Trans id="WelcomePage.toggle_locale">
                 En Español
               </Trans>
             </Link>
           </h1>
-
           <Case test={locationCount}>
             <When value={0}>
               <Trans id="WelcomePage.account_misconfigured">
@@ -111,21 +108,13 @@ export default class WelcomePage extends ReactNComponent<any, State> {
             <When>
               <p>
                 <Trans id="WelcomePage.welcome">
-                  Hi
-                  {' '}
-                  {user.firstName}
-                  ! You've been added
-                  by
-                  {' '}
-                  {this.global.i18n._(plural(this.totalLocations(), { one: 'location', other: 'locations' }))}
-                  {' '}
-                  to Greenlight's secure COVID-19 monitoring platform.
+                  Hi {user.firstName}! You've been added
+                  by {plural(this.totalLocations(), { one: '# location', other: '# locations' })} to Greenlight's secure COVID-19 monitoring platform.
                 </Trans>
               </p>
               <p>
                 <Trans id="WelcomePage.instructions">
                   {this.whoDoYouFillSurveysFor()}
-                  {' '}
                   We will not share any data without your
                   permission.
                 </Trans>
@@ -133,26 +122,23 @@ export default class WelcomePage extends ReactNComponent<any, State> {
             </When>
           </Case>
 
-          <img alt="Welcome to Greenlight!" src="/images/welcome-doctor.svg" />
+          <img alt="Welcome to Greenlight!" src={welcomeDoctorImage} />
           <p>
             <Trans id="WelcomePage.terms_and_conditions">
-              By continuing, you accept Greenlight's
-              {' '}
+              By continuing, you accept Greenlight's{' '}
               <Link
                 onClick={() => {
                   this.setState({ termsOpened: true })
                 }}
-              >
-                Terms and Conditions
-              </Link>
-              .
+              > Terms and Conditions
+              </Link>.
             </Trans>
           </p>
         </Block>
         <Block>
           <Row tag="p">
             <Col tag="span">
-              <Button large onClick={() => signOut()}><Trans id="WelcomePage.sign_out">Sign Out</Trans></Button>
+              <Button large onClick={() => signOut()}><Trans id="Common.sign_out">Sign Out</Trans></Button>
             </Col>
             <Col tag="span">
               <Button
@@ -160,7 +146,7 @@ export default class WelcomePage extends ReactNComponent<any, State> {
                 fill
                 href={paths.welcomeReviewPath}
               >
-                <Trans id="WelcomePage.continue">Continue</Trans>
+                <Trans id="Common.continue">Continue</Trans>
               </Button>
             </Col>
           </Row>
@@ -174,7 +160,7 @@ export default class WelcomePage extends ReactNComponent<any, State> {
           <Toolbar>
             <div className="left" />
             <div className="right">
-              <Link sheetClose><Trans id="WelcomePage.close">Close</Trans></Link>
+              <Link sheetClose><Trans id="Common.close">Close</Trans></Link>
             </div>
           </Toolbar>
           {/*  Scrollable sheet content */}